+++ CHANGES.rst
@@ -3,7 +3,13 @@ Changelog
 
 1.3.1 (unreleased)
 ----------------
-- Fixed a document generation bug when result indicator is None
+- Added new dashboard template for manager follow-up
+  [fngaha]
+- Fix ddetail template
+  [fngaha]
+- Added a presentation option to the follow-up report
+  [fngaha]
+- Fix a document generation bug when result indicator is None
   [fngaha]
 - Enabled categories column on action dashboard
   [fngaha]
@@ -31,7 +37,7 @@ Changelog
   [fngaha]
 - Moved strategicobjective, operationalobjective, pstaction, pstsubaction _fields config from registry to pstprojectspace
   [fngaha]
-- Added plan and plan_values fields and updated ecomptes export
+- Added plan and plan_values fields, updated ecomptes export and templates
   [fngaha]
 - Added optional new style template without numbering
   [bleybaert]